[#129] refactor : parseDate 유틸 함수로 분리
