+++ src/utils/date.ts
@@ -22,3 +22,12 @@ export function getTodayKstKey(): string {
   const kstIso = getNowKSTIsoString(); // ...+09:00
   return kstIso.slice(0, 10); // "YYYY-MM-DD"
 }
+
+// UI 파싱용
+export function getParseDate(iso: string) {
+  const d = new Date(iso);
+  const y = d.getFullYear();
+  const m = String(d.getMonth() + 1).padStart(2, '0');
+  const day = String(d.getDate()).padStart(2, '0');
+  return `${y}.${m}.${day}`;
+}
